+++ thunderstore/CHANGELOG.md
@@ -2,6 +2,7 @@ Hint: this mod is still in Alpha - use at your own risk! These updates can break
 
 ## 26.7.15
 - improvement: searchable settings picker (instead of a simple input field in settings)
+- fix: dungeon loot-multiplicator errored when allowlist/denylist was empty but selected
 - draft: show players & minimap when joining lobby as a client (not as an host)
 - draft: show notifications on the top right instead of bottom right (webinterface)
 - draft: add the ability for custom loading screens (additionally to the dungeon start sounds, yay :D)